feat(server): 角色技能与核心业务逻辑
同步 Character/Client/Skill 及剩余 Java 服务端改动。

+++ gms-server/src/main/java/org/gms/dao/entity/CharactersDO.java
@@ -64,7 +64,7 @@ public class CharactersDO implements Serializable {
 
     private Integer maxmp;
 
-    private Integer meso;
+    private Long meso;
 
     @Column("hpMpUsed")
     private Integer hpMpUsed;
@@ -137,7 +137,7 @@ public class CharactersDO implements Serializable {
 
     private Integer matchcardties;
 
-    private Integer merchantmesos;
+    private Long merchantmesos;
 
     private Boolean hasmerchant;
