+++ gms-server/src/main/java/org/gms/service/GiveService.java
@@ -477,13 +477,13 @@ public class GiveService {
     }
 
     private void doGainMeso(Character chr, int quantity) {
-        int meso = chr.getMeso();
-        long sum = (long) meso + (long) quantity;
+        long meso = chr.getMeso();
+        long sum = meso + (long) quantity;
         if (sum < 0) {
-            quantity = -meso;
+            quantity = (int) -meso;
         }
-        if (sum > Integer.MAX_VALUE) {
-            quantity = Integer.MAX_VALUE - meso;
+        if (sum > Long.MAX_VALUE) {
+            quantity = (int) (Long.MAX_VALUE - meso);
         }
         chr.gainMeso(quantity);
     }
